chore(tidy up): update library to newer tech
BREAKING CHANGE: no longer supports older versions of node or callbacks
